WIP disk failure: basic test coverage for disk failure
This patch adds an EMC test that spawns three servers and triggers EIOs
on two of them to fail two different tablets. With improper
disk-failure-handling, this scenario alone would have been enough to
leave the server with only a single copy of data, as the two servers
with EIOs would have been shut down entirely.

With proper disk-failure handling, this scenario would be salvageable,
and data would be replicated on the remaining disks.

Tests are also added to tests behavior during flushes and scans.

WIP because this test is currently failing on macOS.

Change-Id: I8ff63ec71ab718866484b9f3ec7264bc72ecfe97