disk failure: make DataDirManager failure-aware
The DataDirManager must record what directories are unhealthy in order
to avoid placing new data on failed disks. This patch achieves this by
maintaining a set of UUID indices in the DataDirManager that
correspond to failed directories. Additionally, a count of the number
of known failed directories is maintained as a metric.

Tests are added to data_dirs-test to ensure that failed directories
are not used and are not returned as part of newly created
DataDirGroups. If no healthy directories exist, callers will return an
IOError with posix code ENODEV.

Change-Id: Iee212793152de5de5198751d649ab34fb97f6aa2