disk failure: test disk failures during flushes
This patch adds a test to test the failure of flushes (MRS and DMS) on a
server.

Testing of MRS flushes is accomplished by causing disk failures on a
single data dir of a single server, allowing the cluster to re-replicate
the data, failing the new data dir, and repeating.

Testing of DMS flushes is accomplished by upserting the same range of
rows multiple times while causing disk failures. This should cause a
server to crash and recover.

Change-Id: Ibe16729c9d3836949fda56b98854051b6c83c926