tests: IOErrors and IllegalState in TestWorkload
A TestWorkload has the ability to expect various errors during its run.
It can now expect IOErrors or IllegalState during a scan. This is
helpful in testing disk failure during scans.

An IOError may surface if the disk with tablet data fails in the middle
of a scan. An IllegalState may surface if the disk failed and the tablet
was failed before the scan starts.

Change-Id: I364c0ae2ac48920bcbd5b662b931ca448464c90e