mini_tablet_server: support multiple data dirs
This patch gives MiniTabletServers the ability to start with multiple
data dirs. A number of dirs is specified and directory names are
generated for the specified number. Additionally, a WAL dir will be
generated with an appropriate suffix.

The original directory structure is kept as a default.

E.g. if the number of data dirs is 3, the following directories will
be generated:
/test_dir/test_tserver_wal_dir
/test_dir/test_tserver_data_dir-0
/test_dir/test_tserver_data_dir-1
/test_dir/test_tserver_data_dir-2

Original:
/test_dir/test_tserver

A test is added to tablet_server-test to exercise this.

Change-Id: I52c9352f1a3565d58149cf2c63d37246c6b39c23