WIP disk failure: add persistent disk states
This patch adds a new version of the path set so failed disks are not
used the next time the server is run. This is needed, as data on a
failed disk may be partially written and thus should not be used.

To accomplish this, disk states, path names, and a timestamp are added
to the path set. Additionally, if any disks are failed when loading
instances from disk, an 'invalid' instance with no path set metadata
will be created instead of failing the startup process.

To localize disk failure handling at startup, disk failures are ignored
until the DataDirManager and BlockManager are initialized in
FsManager::Open().

CheckIntegrity() is updated to accomodate this. Rather than
comparing all instances against an agreed-upon set of UUIDs, the single
path set with the largest timestamp is used as the main one to compare
against (if only old path sets are available, the first healthy one is
used). If no instances are healthy, CheckIntegrity() will fail, as all
disks are failed.

The main path set is checked to ensure its integrity (e.g. no
duplicates), after which it is upgraded with the extra metadata if
needed. It is then checked to ensure it is consistent with the healthy
instances in terms of having the right UUIDs and paths.

Testing is done in three ways:
- unit testing for the new iteration of CheckIntegrity()
- testing in ts_disk_failure-test to ensure failed disks will not fail
  when starting up the server
- testing in disk_failure-itest to ensure that starting up with a failed
  disk will cause the tablets on the failed disk to replicate

Some notes:
- In the case of a server restart where all healthy disks fail to start
  up and some known failed disks start working again, the server will
  successfully start up with the bad disks (may have partially-written
  data).
- If there are any unhealthy instances when upgrading the path sets (i.e.
  adding disk states, paths, timestamp), a complete mapping of UUIDs to
  paths will not be available, and CheckIntegrity() will fail with an
  IOError.
- The main path set's disk states are updated to reflect the failure of
  the instances. Since data on a failed disk cannot be trusted, disks
  that are successfully read from but are already marked FAILED by the
  main path set are not marked HEALTHY.

WIP because I'd like to add more tests (will add more when other disk
failure patches stabilize). The contents are reviewable and I'd
appreciate feedback on the new integrity checks.

Change-Id: Ifddf0817fe1a82044077f5544c400c88de20769f