disk failure: replicate failed tablets
Tablets that are marked FAILED should eventually be replicated.
A tablet will be FAILED if it fails to bootstrap or start.

This patch also adds a new tablet state called FAILED_AND_SHUTDOWN for
cases where the tablet needs to be shutdown (e.g. to stop responding to
heartbeats).

Testing is done in itests by failing and shutting down a tablet, and
ensuring it is replicated.

Change-Id: I5f61585b02fbe270d215bf7f49c0d390ceee3345