disk failure: forced shutdown of MvccManager
When a disk fails, tablets with data on the failed disk must shut down.
Normally, when a replica is shutdown, it must wait for all transactions
to finish. However, transactions that fail due to disk failure will not
complete.

This patch adds the ability to Cancel transactions. Unlike Aborts, which
cannot end a transaction that is in the APPLYING state, Cancels can end
any transaction. This is only "safe" because the tablet to whom the
MvccManager belongs is expected to shutdown.

Testing is done by canceling all transactions for a tablet's MvccManager and
ensuring that the manager can no longer be used.

Change-Id: I983620f27e7226806a2cca253db7619731914d42