WIP disk failure: survive disk failures
Disk failures are a special case of errors that can be handled. Certain
code paths pass along disk failure Statuses until they eventually hit a
CHECK and crash Kudu. With this patch, these code paths will instead
allow Kudu to continue running, under the assumption that the errors are
handled.

Errors are handled by synchronously failing the tablets on the failed
disk, and asynchronously shutting them down.
WIP: some asynchronously-run tasks need to be taken into account:
- maintenance ops that may still be in the process of running need to be
  aborted and unregistered
- tablet copies that are currently running should be updated such that
  one that copies to a failed disk will not complete bootstrapping
- transactions that are currently in flight need to be safely aborted
  - this is somewhat handled with txn.Cancel()

A small test is added to tablet-test to insert some data and fail.
Rather than crashing, the end-state of the tablet must indicated a disk
failure.

This is a part of a series of patches to handle disk failure. To see how
this patch fits in, see section 2.4 of:
https://docs.google.com/document/d/1zZk-vb_ETKUuePcZ9ZqoSK2oPvAAaEV1sjDXes8Pxgk/edit

Change-Id: I109635a54268b9db741b2ae9ea3e9f1fe072d0a8