WIP disk failure: eject failed txs from tracker
Currently, if a TransactionState object is deleted while it is Applying,
Kudu will crash as a way to prevent the failed transaction from
persisting incorrect state onto disk. However, this is not always
desirable; transactions that fail due to disk failure can keep the
server alive by marking the tablet failed (and unusable). In this case,
there needs to be a way to release transactions from the
TransactionTracker without crashing Kudu, regardless of state.

This patch adds the ability to Cancel transactions. Unlike Aborts, which
ensure the transaction is not in the APPLYING state (must be RESERVED)
when destructed, a Canceled transaction has no such constraint. This is
only "safe" because the TransactionTracker's tablet is expected to
shut down and the failed disk will no longer be used.

WIP because there is no testing and I'm not sure this is the best design
for what I'm trying to accomplish.

This is part of a series of patches to address disk failure. To see how
this patch fits in, see section 2.3 of:
https://docs.google.com/document/d/1zZk-vb_ETKUuePcZ9ZqoSK2oPvAAaEV1sjDXes8Pxgk/edit

Change-Id: I983620f27e7226806a2cca253db7619731914d42