KUDU-1489: move tablet metadata
The crux of this patch is aimed at moving metadata when initializing the
TSTabletManager. Rather than always getting metadata from the first data
directory, all known directories (both WAL and data dirs) are scanned
for metadata instances. Any metadata found in a directory it shouldn't be
in is moved to one appropriate for it.

An appropriate directory for a tablet's metadata is:
1. any directory in the tablet's directory group, or
2. the WAL dir (if tablet metadata striping is disabled)

If striping is enabled, the metadata directory for each tablet will be
decided when the tablet is created (or copied).

Change-Id: Ia5308f8d4b4b738f353f511e2f1a2634b675a60a