mvcc: allow tablet shutdown without completing txs
Currently, the only way to stop an Applying transaction is to wait for
it to finish and Commit it. This constraint was put in place to
guarantee on-disk correctness, but is sometimes too strict. E.g. if the
tablet is shutting down, the Apply doesn't need to finish.

This patch adds the ability to "stop" a tablet by closing its MVCC
manager. Once this happens, new Applies will return and not move to the
Commit phase, and any methods waiting for the tablet's Applies to Commit
(e.g. new snapshot scans, FlushMRS) will respond with an error
immediately. Applies that are already underway may still Commit, but
these Committed operations are inconsequential w.r.t. consistency;
having some in-flight transactions Commit and others not is consistent
with the server crashing in between the Commits of two transactions.
Additionally, once the MVCC manager is closed, new transactions will
abort immediately before even reaching the Prepare phase.

This will be particularly useful in handling disk failures: if a tablet
needs to be shut down due to disk failure, its MVCC manager can be
closed immediately, allowing currently-Applying transactions to abort
without Committing.

This patch only includes the behavior when shutting down a tablet, with
the assumption that a tablet will only be shut down when it's being
deleted and we don't care too much about its in-flight transactions
Committing. Code paths that previously crashed Kudu if Applies did not
succeed will now not crash if the MVCC manager is closed and log a
warning instead.

Testing is done by adding the following:
- a test in mvcc-test to shut down MVCC and delete an Applying
  transaction, ensuring that there are no errors when it leaves scope.
- a test in mvcc-test to wait on an Applying transaction, shut down
  MVCC, and ensure that any waiters will return with an error.
- tests in ts_tablet_manager-itest that ensure stopped leaders don't
  complete writes and stopped followers do.

Change-Id: I983620f27e7226806a2cca253db7619731914d42