itest: allow use of verifiers with EMCs that specify non-default block managers
ClusterVerifier and LogVerifier both rely on opening a new FsManager,
using the current FLAGS_block_manager of the caller process.
ExternalMiniClusters, on the other hand, grant the flexibility to
specify any flags, and create the cluster in a separate processes.

This precluded the Verifiers from operating on ExternalMiniClusters that
are specified with different block manager types than that of the
Verifier's caller (e.g. in parameterized EMC tests).

To remediate this, block manager type has been added to various
constructor Options (i.e. ExternalMiniClusterOptions,
ExternalDaemonOptions, FsManagerOpts, DataDirManagerOptions).
This field is used by the Verifiers when opening a FsManager.

There is also some cleanup of ExternalDaemon to store and use
ExternalDaemonOptions instead of the individual members.

To test, I parameterized disk_failure-itest, which exercises using a
ClusterVerifier and LogVerifier, and restarts an ExternalTabletServer.

Change-Id: Iaa44eb33d0c025830f97a2ed7583c8186f915e94