KUDU-1613: evict replicas with wrong server UUIDs
Currently, if a tablet server is rebuilt before the replicas on it are
evicted, they won't ever be evicted. With this patch, upon receiving a
response with WRONG_SERVER_UUID, leaders will treat the follower as
failed and handle it appropriately.

A new test case is added to raft_consensus-itest that is set up to
prompt eviction (i.e. low timeouts, ample servers to replicated to), and
without the change, it fails.

Change-Id: If0c7ae9dfbd99c2c9a38e8b1f7748c70882e06b4