internal_mini_cluster: support Cluster/LogVerifier
This patch encapsulates MiniClusterFsInspector into an abstract class
introduces the subclasses it as the existing External- and a new
InternalMiniClusterFsInspector.

With these classes, the LogVerifier, and thus, the ClusterVerifier
can support both External- and InternalMiniClusters.

To test, I updated a test case in ts_tablet_manager-itest to make use of
the new ClusterVerifier with an internal cluster. CheckCluster() uses a
LogVerifier, which in this case uses an InternalMiniClusterFsInspector.

Change-Id: I228a6e3ba1a42db4e243ffdc5116f0c60ee04a84