tablet: add tmeta manager
In the same spirit of decoupling metadata from their underlying storage,
this patch follows ConsensusMetadataManager's lead by implementing a
TabletMetadataManager. All methods in TabletMetadata that currently
touch files now do so through the TabletMetadataManager.

Note: There was a state machine that I removed from of the TabletMetadata
class that tracked on-disk state.

Change-Id: If664082cca3ae2e20f74364a15533d13f99a7757