[no review] exacerbate race between consensus and prepare thread
RaftConsensusElectionITest.ChurnyElections would fail ~5/1000 runs in
DEBUG mode before due to a race between consensus's newfound ability to
advance the MVCC safe time and the tablet's initialization of
transactions.

Even though we check leadership before submitting the transaction to the
prepare thread, it's possible that the leadership can change while the
transaction is being executed, at which point a no-op may advance the
MVCC safe time past a transaction that is about to start.

0. A* checks that it is the leader of term N
1. A* assigns timestamp 999 to a transaction in term N
2. A steps down and B* becomes leader of term N+1, asserting no-op with
   a timestamp 1000
3. A advances its safe time with no-op to 1000
4. A _starts_ the transaction at timestamp 999 but hits a check failure
   because 999 < 1000

I exacerbated the race by sleeping after step 1. A run can be found:
http://dist-test.cloudera.org/job?job_id=awong.1537503850.62520

It's worth noting that the race doesn't wind up on disk because before A
can the write transaction with timestamp 999 to disk, it will check to
see if it's still leader.

Change-Id: Icb07e00e5d3bae0d03d8d263118a6729c0917504