[wip] KUDU-2971 p3: concurrent requests support for subprocess
This patch adds a subprocess server to support multiple requests at a
time. The subprocess client can post request messages to a blocking
queue and a dedicated thread in the server waiting on the queue and
send the message to the underlying subprocess once getting one. The
server has another dedicated thread blocks on receiving messages from
the subprocess, and notifies the client the response with the matching
message ID.

TODO: add more tests for the server, add a dummy client for test, add
retry logic in the client.

Change-Id: I6fddbc389d8c3582bda301b9390700068e3806b2