wip: KUDU-3011: quiesce scans, quiesce leadership
Metric for the number of leaders on a tablet server.
- TODO: maybe redo plumbing?
Test for quiescing leadership.

This plumbs a quiescing state to the tablet server. With this state,
a tablet server will stop:
- admitting new scans
- calling elections

In general the goal is for the tablet server to stop signing up for
doing new work so we can (with some help) seamlessly quiesce and shut it
down.

end-to-end, test a rolling restart
TODO: count the number of leaders

Change-Id: I6700f7d86eea886ec5ad819d2faa4ba6e5080e0e