docs: add a blurb about RF=1 tables when doing a rolling restart
The documented steps don't mention anything about how a RF=1 table will
not make progress in quiescing. We don't expect such tables to be very
common; nevertheless it'd be nice to at least mention the limitation.

While not yet easily automated since there isn't a convenient way to get
the tables that have RF=1, this adds a warning and a recommendation of
tooling that might help.

Change-Id: I2d43d0331dc6d67a9ccb9e84c38b302eed3ae6a6